requirements.txt
jieba==0.42.1
matplotlib==3.7.1
numpy==1.24.2
onnxruntime==1.15.1
opencv_python==4.5.5.64
pandas==2.0.3
Pillow==9.5.0
Pillow==10.0.1
PyAutoGUI==0.9.54
pyclipper==1.3.0.post4
pyshortcuts==1.9.0
pywin32==306
Shapely==2.0.1
streamlit==1.25.0
